internapi/requirements.txt
aniso8601==6.0.0
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-haystack==2.8.1
djangorestframework==3.9.4
drf-haystack==1.8.5
graphene==2.1.6
graphene-django==2.3.2
graphql-core==2.2
graphql-relay==0.4.5
Markdown==3.1.1
promise==2.2.1
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
